I fixed the bug
So, technical explanation on how it was fixed, what I did, and what we were trying to do. Skip to the next paragraph to hear me rant about how much I hate this project:
Problem was that the reservations were being passed to the page as a whole. Meaning all the reservations from all the users was being displayed to all users. That is obviously wrong! What me and Eric tried to do for fucking 2 hours was remodelling the whole dataflow of our database because we thought that the fix can be done in the actual web page... FUCKING WRONG, THAT LOGIC IS FUCKING WRONG. VERY STUPID LOGIC. I WISH I HAD MORE THAN 5 BRAIN CELLS TO REALIZE HOW STUPID OF AN IDEA THAT WAS FOR ME TO TRY TO ACCOMPLISH, AFTER WRITING SPAGETTI CODE THAT EVEN MICROSOFT DEVELOPERS WONT UNDERSTAND. The solution was really simple, instead of doing this:

    return View(new Wrapper(_context.Reservations.ToList(), _context.Restaurants.ToList()));

You should be doing THIS:

    return View(new Wrapper(_context.Reservations.Where(r => r.CustId == user.Id).ToList(), _context.Restaurants.ToList()));

And to fix the user.Id bullshit, you do this fucking grabage, change this:

    public int CustId { get; set; }

TO FUCKING THIS:

    public string CustId { get; set; }

THIS WAS THE FIX, I WILL NEVER GET THOSE 2 HOURS BACK AGAIN

So yeah now it is fixed
But it costed both my mental health, and Erics. Me and him spent 2 hours, trying to find the silly little bug. The bug what we thought would be the MOST SIMPLE THING TO FIX IN THIS PROJECT. Nah, we jinxed the living shit out of that. It got to the point where Eric actually started screaming with me. I HAVE NEVER SEEN HIM EVER DO THAT! IT TAKES A LOT FOR THAT MAN TO BE VISIBLY UPSET, AND BOY OH BOY, WAS HE UPSET! Programming has never been this frusterating. This project made me go insane several fucking times, I am not even exaggerating. When I found out that the bug was fixable by just changing 2 lines, I was furious, enraged, and disapointed. This whole fucking project has fucked my life up.Never will I ever wake up and say "Hmm, lets make a cool website in ASP.NET"
NEVER